Adds build step using vite and breaks code into JS6 modules (#14)
This is a large update. I've selected Vite as the build system for this project. I tried other tools like golang minify but this had the most useful features and serves as a dev server too. I think a lot can be done to shore up the management of global state in the application. For now this works.

closes #13

#### Changes
- adds vite for dev/build
- Breaks out JS into JS6 modules
- Updates docker file to have two stage build

Co-authored-by: Drew Bednar <drew@androiddrew.com>
Reviewed-on: https://git.runcible.io/androiddrew/webserial/pulls/14